src/search/requirements.txt
grpcio
grpcio-tools
requests